+++ src/main/java/com/shinesolutions/aemorchestrator/actions/Action.java
@@ -1,6 +1,6 @@
 package com.shinesolutions.aemorchestrator.actions;
 
-public interface ScaleAction {
+public interface Action {
 
     boolean execute(String instanceId);