+++ src/main/java/com/shinesolutions/aemorchestrator/exception/MessageHandlerNotFound.java
@@ -1,0 +1,11 @@
+package com.shinesolutions.aemorchestrator.exception;
+
+public class MessageHandlerNotFound extends Exception {
+
+    private static final long serialVersionUID = 1L;
+
+    public MessageHandlerNotFound(String messageSubject) {
+        super("No message handler found for message with subject: " + messageSubject);
+    }
+
+}