+++ src/main/java/com/shinesolutions/aemorchestrator/handler/MessageHandler.java
@@ -1,9 +1,7 @@
 package com.shinesolutions.aemorchestrator.handler;
 
-import javax.jms.Message;
-
 public interface MessageHandler {
-    
-    boolean handleMessage(Message message);
-    
+
+    boolean handleEvent(String message);
+
 }